Style header and body with CSS
This commit styles the header and body elements with CSS, and creates classes in the header element to better aid that styling. 

The following classes were created:
- .logo
- .header-links

The .logo class includes the following CSS properties:
- font-size: 24px;
- font-weight: 700;
- color: #F9FAF8;

The .header-links class includes the following CSS properties:
- font-size: 18px;
- color: #E5E7EB;
- display: flex;
- gap: 25px;

The following CSS properties were added to the body element:
- margin: 0;
- display: flex;
- flex-direction: column;

The following CSS properties were added to the header element:
- background-color: #1F2937;
- padding: 200px;
- padding-top: 25px;
- padding-bottom: 25px;

The following CSS properties were added to the ul element nested within the header element:
- list-style: none;
- margin: 0;
- padding: 0;
- display: flex;
- justify-content: space-between;
- align-items: center;

These changes were made to implement the design mockup of the header provided in the Landing Page project's design image, and ensure future sections of the website stack as shown in the design image.