Style section 2 of page content with CSS
This commit styles section 2 of the page's content with CSS.

The following CSS properties were added to the .section-2 class:
- padding: 200px;
- padding-top: 25px;
- padding-bottom: 25px;

The following CSS properties were added to the .section-header class:
- text-align: center;
- font-size: 36px;
- font-weight: 800;
- color: #1F2937;

The following CSS properties were added to the .card-container class:
- display: flex;
- padding-top: 20px;
- justify-content: center;
- gap: 20px;

The following CSS properties were added to the .card class:
- display: flex;
- flex-direction: column;
- gap: 10px;

The following CSS property was added to the .text class nested within the .card class:
- font-size: 18px;

These properties were added to ensure the cards rest below the section header, and are displayed horizontally. It also ensures the section header is displayed from the middle instead of the line-start, and the font sizing and padding adhere to the design mockup presented in the Land Page project's design image.