Adjust padding and indent quote for section 3
This commit adjusts the padding for the third section of the page content to bring it inline with section 2. It also introduces padding to the left of the quote to indent it.

The following CSS property value was adjusted for the .section-3 class:
- padding: 325px; to padding: 275px;

The following CSS property was also introduced to the .section-3 class:
- padding-left: 40px;

These changes were made to resemble the design mockup in the Landing Page project design image.