Style section 4 of page content with CSS
This commit styles the fourth section of the page content with CSS.

The following CSS properties were added to the .section-4 class:
display: flex;
- justify-content: space-between;
- background-color: #3882F6;
- margin: 200px;
- margin-top: 75px;
- margin-bottom: 75px;
- padding: 150px;
- padding-top: 50px;
- padding-bottom: 50px;
- border-radius: 4px;

The following CSS properties were added to the .main-text class:
- color: #FFFFFF;
- font-size: 24px;
- font-weight: 700;

The following CSS properties were added to the .secondary-text class:
- color: #FFFFFF;
- font-size: 18px;

The following CSS properties were added to the button element nested within the .section-4 class:
- color: #FFFFFF;
- background-color: #3882F6;
- border: solid 2px #FFFFFF;
- border-radius: 4px;
- padding: 2px;
- padding-left: 15px;
- padding-right: 15px;

These additions ensure section 4 is a blue box surrounded by a white background. Its corners are rounded, and the information is on the left side of the box, and the sign up button is on the right. This resembles the Landing Page project's design image.