Center placeholder image in section-1
This commit centers the placeholder image in the first section of the
page content when resizing the window's width.

The following HTML was added to accomidate CSS properties to be added
later:
<div class="hero-image">
    <div class="placeholder">this is a placeholder for an image</div>
</div>

The following CSS rule was created to center the placeholder image when
the window's width is being resized:
.hero-image {
  display: flex;
  flex: 1;
  justify-content: center;
  align-items: center;
}

This change was made so the image can be inline with the other content
on the webpage if the user resizes the window's width.