Rename various classes
This commit renames various classes to aid in readability.

The following class names were changed in section 1:
- info -> hero-left
- hero-text -> hero-main-text
- text -> hero-secondary-text
- hero-image -> hero-right
- placeholder -> hero-image

The following class was created in section 1:
- hero-sign-up

The following class names were changed in section 2:
- placeholder -> card-image
- text -> card-text

The following class names were changed in section 4:
- info -> call-left
- main-text -> call-main-text
- secondary-text -> call-secondary-text

The following class was created in section 4:
- call-sign-up

The following class name was changed in the footer:
- text -> footer-text

These changes were made to specify the intended function of each class.
Previous class names were more generic, and the new class names are
more specififc.